requirements.txt
kafka-python==2.0.2
yfinance==0.2.31
pandas==2.1.1
pytest==7.4.2
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0